dristi_adk_agent/requirements.txt
google-adk 
google-cloud-vision 
google-cloud-bigquery 
googlemaps 
fastapi 
uvicorn
google-cloud-videointelligence 
google-cloud-bigquery